Scriptable Object Issue
한 스크립트에 다 뭉쳐놔서 오류 발생
파일 분리해서 해결 완료.
