+++ Assets/Resources/ScriptableObj.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2de98aa441fe68040b0337d976ee8b80
+guid: 6da8c6f290ae0f44599787178c9eacb7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
