+++ Assets/Scripts/Ui/Gacha/GachaCookie.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7ba540adc56db0b459d63a4f0b58fbee
