+++ Assets/Scripts/Ui/Gacha/GachaGear.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "GachaGear", menuName = "Gacha/Gear")]
+public class GachaGear : ScriptableObject
+{
+    public string itemName;
+    public string itemId;
+    public float weight;
+    public Sprite Icon;
+}
