+++ Assets/Scripts/Ui/Gacha/GachaGear.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c8857a0f44d04e14ea5388d14ed3a7b9
