+++ Assets/Scripts/Ui/Gacha/GachaManager.cs
@@ -3,28 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine.UI;
 
-
-[CreateAssetMenu(fileName = "GachaGear", menuName = "Gacha/Gear")]
-public class GachaGear : ScriptableObject
-{
-    public string itemName;
-    public string itemId;
-    public float weight;
-    public Sprite Icon;
-}
-
-[CreateAssetMenu(fileName = "GachaCookie", menuName = "Gacha/Cookie")]
-public class GachaCookie : ScriptableObject
-{
-    public string cookieName;
-    public string cookieId;
-    public float weight;
-    public Sprite Icon;
-
-    // 스킬 스크립트도 여기 추가해서 써도 됌
-}
-
-
 public class GachaManager : MonoBehaviour
 {
     private float totalWeight;
